requirements.txt
streamlit==1.36.0
google-generativeai==0.7.1
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.4.0 # Added for safety, good practice with Streamlit for image handling if ever needed